requirements.txt
mcp
mcp[cli]
uv
openai
ollama
python-dotenv
anthropic
httpx
argparse
requests
